+++ server/src/Domain/Events/InvalidOrderSummaryNoticedEvent.cs
@@ -1,3 +1,3 @@
 namespace EveBuyback.Domain;
 
-public record InvalidOrderSummaryNoticedEvent(ItemType Item, int Volume);
+public record InvalidOrderSummaryNoticedEvent(ItemType Item, long Volume);
