+++ server/src/Domain/Models/VerifiedContractItem.cs
@@ -2,4 +2,4 @@ namespace EveBuyback.Domain;
 
 public record VerifiedContractItem(
     ItemType Item,
-    int Volume) : ContractItem(Item.Name, Volume);
+    long Volume) : ContractItem(Item.Name, Volume);
